_ALTRIX_-main/ai-humanizer/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pdfplumber
pytesseract
Pillow
torch
transformers
